atualizacoes dia 01/04/2026

+++ src/main/java/com/senai/conta_bancaria/application/service/UsuarioService.java
@@ -3,6 +3,7 @@ package com.senai.conta_bancaria.application.service;
 import com.senai.conta_bancaria.application.dto.UsuarioRequestDTO;
 import com.senai.conta_bancaria.application.dto.UsuarioResponseDTO;
 import com.senai.conta_bancaria.domain.entity.Usuario;
+import com.senai.conta_bancaria.domain.exception.UsuarioNaoEncontradoException;
 import com.senai.conta_bancaria.domain.repository.UsuarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -29,20 +30,25 @@ public class UsuarioService {
     //Metodo para buscar usuario especifico
     public UsuarioResponseDTO buscarUsuarioPorId(Long id) {
 
-        return UsuarioResponseDTO.fromEntity(usuarioRepository.findById(id).get());
+        return UsuarioResponseDTO.fromEntity(usuarioRepository.findById(id).orElseThrow(() -> new UsuarioNaoEncontradoException(id)));
     }
 
     public UsuarioResponseDTO atualizarUsuario(Long id, UsuarioRequestDTO usuarioRequestDTO) {
-        Usuario usuarioAtualizado = usuarioRepository.findById(id).get();
+        Usuario usuarioAtualizado = usuarioRepository.findById(id)
+                .orElseThrow(() -> new UsuarioNaoEncontradoException(id));
 
         usuarioAtualizado.setNome(usuarioRequestDTO.nome());
         usuarioAtualizado.setEmail(usuarioRequestDTO.email());
         usuarioAtualizado.setSenha(usuarioRequestDTO.senha());
         return UsuarioResponseDTO.fromEntity(usuarioRepository.save(usuarioAtualizado));
 
+
     }
 
     public void deletarUsuario(Long id) {
+        if (!usuarioRepository.existsById(id)) {
+            throw new UsuarioNaoEncontradoException(id);
+        }
         usuarioRepository.deleteById(id);
     }
 }
